AdminPolicy now subclasses ActiveFedora::Base (fixes #1000) Common index-related behavior factored out of DulHydra::Base into DulHydra::Indexing
