+++ spec/factories/admin_policy_factories.rb
@@ -2,7 +2,7 @@ FactoryGirl.define do
 
   factory :admin_policy do
     # AdminPolicy title must be present and unique
-    sequence(:title) { |n| "Admin Policy #{n}" }
+    sequence(:title) { |n| ["Admin Policy #{n}"] }
 
     factory :public_read_policy do
       default_permissions [{:name => "public", :type => "group", :access => "read"}]
